Begin with architecture refactor
Instead of having manage be a loop that listens for changes to a queue modified with exposed methods GetBuffer and PutElem, instead write into a (buffered) channel that Manage() listens to. The data that needs to be written is stored in the channel. Execution is then halted in the main loop when using the Algobuffer. Need to think more about this.