Remove unneeded Buffer interface
At this time, the Buffer interface doesn't seem to provide value.
This is because the AlgorandBuffer requires graceful initialization and
shutdown, something that a normal buffer wouldn't normally need.

The abstraction feels forced. As such, I believe it's better to keep the
AlgorandBuffer as-is, and having the Oracle know about its existence.

This also makes sense when thinking about the value of this Oracle
system. Ideally, I want to push ALL KINDS of data in a similar way to
the blockchain. The oracle is essentially a bridge connecting public
data feed APIs to the Algorand application. What the user needs to write
is a translation module that translates raw data into a
map[string]string format, with the appropriate limitations. The Oracle
struct can then expose the necessary abstractions to allow for these
translation modules. The core Oracle fetch-and-write cycle would largely
be unchanged.

If the Oracle reaches this point, I will most likely transfer it to the
Siam project and have this repository here host only the translation
module and configuration for the Oracle. This kind of structure would
definitely be fit to support these future use-cases.